Hops testing
passes if nick names don't screw it up.
somehow this works most of the time but somtimes(TM) it failes.

In these cases the nicknames are on the path instead of the refs.
Maybe the debuginfo somehow corrupts the graph?!

Move Hops to Builder:

wanted to keep this general ontop of Graph but couldn't figure out the
walker code so just made it work with some recursive decent in the
meantime...

Leaving gossip.ping open:

if we close it properly we crash some clients.
if we close it with an error some peer think we failed the test.
I don't want to break either case so i will just leave the stream open.

also
* rm uneeded debug info
* close test store
* use hops instead of follows
* disable logbuilder tests
* generalize gossip fetcher code
* extra activeFetch lock
* fix sbot compilation
* actually call cancel
* move up timeout - conn timeout is 15m so give it some time