node: Add integration test
Add an integration test for the linux network stack using network
namespaces.  Test against a small set of IPv4 subnets, and IPv6, and
confirm the received advertisement.

For the test to work, changes to the SUDO file are suggested within the
integration test file.

Include, but comment out, a breaking test which requires fixing.