rename node to network
I think _node_ is a confusing enough term as it is in p2p stuff.
additionally so since the ref implementation of ssb is in nodejs.

move test code around and use discoverer:
the defered cleanups (to remove the netns and bridge) needs manual
intervention when a test fails (name exists on next run) so i moved the
remove before the create to just wipe the previous run.

also: split advertiser' listen portion into discoverer

on the multi-network test code:
i turned the cmd.Run of the sending into returning the cmd.

i start and wait it in a goroutine and sync it with the receive

this way there is only the sleep in TestSendAdvertisement and it's
network is not killed by the next test case before it's done.