small fixes
* shallow fsck
* some compilation fixes
* supress and bail out of cancel error flood
* go mod vendor muxrpc@cleanup
* blobs: get then want
* denoise log output
* skip invalid but show key
* sbot: don't panic on no-network mode
* feed sync test
* get to same base before check (fixes #14)
* without ptrs the locks dont work
* fix Hops (updates #4)

* network: change to lastWinsTracker
works much better with stalled connections from the js world.
so it's better ot have this for the pubs by default.

TODO: make configurable and have multiple listeners so that multiple
local client connections work

* remove hasOwn pull
this is a relic from trying to re-sync existing feeds.
currently this is part of the hops set anyway.

* cleanup gossip.ping once more

"just leave it open" didn't work.
there was a defered close in HandleCall.