remove vendor/
this was mostly a go1.10 thing, though `go mod vendor` is also _usefull_
to ship patches where you don't want to maintain a fork repo but this is
really not nice practice.  all hail go.mod !

additionally, vendor/ was only intended for main projects and this leads
to errors like `type A cant be used in place of (vendor/path).typeA`
which can be annoying to understand. (TODO: drop vendor/ from muxrpc and
secretstream)