improve connection wrapping
before we only had conn wrapping that happend after the shs+boxstream
was done. The new version of netwrap allows multiple wrappers and thus
allows us to wrapp before the connection was authenticated and
encrypted.

This is primarily usefull to apply options on the syscall.Conn (like
socket fd options).