add 3rd bot to reindexing case
this showed a problem with the bitmap indexes where the order is not
kept as (previous) messages are added. ie first message is A and the set is [A].

As a message B which is before A is discovered (decrypted) it get's added to the
same set but since its receive sequence is smaller then A the set looks
like [B, A] and reading the 2nd message from that set will result
in processing A twice.