start dropping pkg/errors
it feels like the nicer API but talking a stack-trace every time
introduces quite the performance penalty. Especially on the semi-happy
paths where error values are just signals,
like ErrNoSuchBlob or EndOfStream.

mostly done except some cmd tools and some _test code where it's used as
check(errors.Wrap(err, "the msg"))
which uses the cases that errors.Wrap(err) returns nil if err is nil

replacing it there would need a lot of if err!=nil { ... "the msg "}
also, that is just test code and not performance critical