fix bug when listing empty mf + creating indexes
before this patch, indexes were being created on the root metafeed,
instead of the `indexes` subfeed.

this patch remedies that problem, and tackles the follow-on effects of
needing to regard the `indexes` feed when performing actions on an index
feed.

another bug was also found where we would return an error when trying to
list an empty metafeed. this was fixed by using the
`keys.IsNoSuchKey(error)` function call to detect the case & return an
empty slice if detected.