Always replicate your own feed
It appears that go-ssb has a mechanism for replicating your own feed
which is supposed to be triggered if the local database is removed.
Unfortunately the current implementation of this mechanism has two
problems:
- it appears that if the feed wasn't fully replicated then the
  replication will not be resumed unless it manages to fully complete
  during the first attempt
- this mechanism will never be triggered if an authoriser which
  authorises all connections (always returns nil) is set through
  WithPublicAuthorizer

To mitigate this I ensured that the local feed will always be present in
the list of feeds to be replicated.

An additional bug that I discovered appears to prevent the list of
wanted feeds to be updated unless the receive log changes at least once.
This may be a desired behaviour however it prevented me from placing a
line with my fix within the update method of graphReplicator.