Fix a "deadlock" in WantManager
The code establishes a lock which is locked for up to 3 * n_peers
minutes. My fix attempts to release the lock early.

We may consider not querying other peers at all in the future as
presumably we should get separate has messages from them. In general I
don't understand why this fallback code querying other peers is needed.

I removed the code which removes a blob from the wants list if we can't
retrieve it in one go. I don't understand the purpose of this code as I
think that the download of a blob should be reattempted.