Fix incorrect field reuse in LegacyGossip (#149)
StartLegacyFetching accepts a parameter withLive to control whether live
streaming is used. At the same time field enableLiveStreaming is used to
control the behaviour in the HandleConnect method.

On top of all of that replicateNegotiator specifies a separate flag for
this behaviour. This flag is never set. Instead of this flag the flag
from LegacyGossip should be used.

To address all of this I removed the flag from the method completely and
now the implementation relies on the field in LegacyGossip. I completely
removed the flag from replicationNegotiator as it is not set. I also
removed the other flag disableEbt which is never used and just adds to
the confusion. Disabling ebts is handled differently by the code.